smoothed genre tables
Data looked jagged -I think our raw data records all revenue to the movie by the month it was released, not by monthly ticket sales.  To try and mitigate this a bit, I created a table of revenue type by genre using 6 month averages.  This is stored in smoothed_genre_share.csv, and the excel I used to create this, with the pivot table and smoothing tables, is smoothed_genres.xlsx.

Ideally would want to do the smoothing in Python or VBA, so we could change the number of months to smooth; will leave as a possible future enhancement.